refactor: Extract ContentViewModel into separate ViewModels.
