+++ app/src/main/java/no/usn/mob3000/ui/screens/info/faq/CreateFAQScreen.kt
@@ -3,7 +3,6 @@ package no.usn.mob3000.ui.screens.info.faq
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
-import no.usn.mob3000.domain.viewmodel.ContentViewModel
 import no.usn.mob3000.ui.components.info.ContentEditor
 
 /**
