+++ app/src/main/java/no/usn/mob3000/ui/screens/info/news/NewsScreen.kt
@@ -12,7 +12,6 @@ import androidx.compose.runtime.*
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.StateFlow
 import no.usn.mob3000.domain.model.NewsData
-import no.usn.mob3000.domain.viewmodel.ContentViewModel
 import no.usn.mob3000.domain.viewmodel.LoginState
 import no.usn.mob3000.ui.components.info.ContentItem
 import no.usn.mob3000.ui.components.info.PaddedLazyColumn
